--- v2/internal/x509util/pool.go
@@ -1,13 +1,0 @@
-package x509util
-
-import (
-	"crypto/x509"
-)
-
-func NewCertPool(certs []*x509.Certificate) *x509.CertPool {
-	pool := x509.NewCertPool()
-	for _, cert := range certs {
-		pool.AddCert(cert)
-	}
-	return pool
-}
